Added support for command line flags and arguments.
Command line and HTTP interfaces of the debugger can now be selectively turned
on and off by the use of command line flags: --cmdline, --nohttp.

The docroot of the code to debug can also be specified using the --docroot flag
and its argument.  By default, we assume that this docroot is the same as the
current working directory.