Update source code in UI.
A new "refresh" icon is displayed *before* filenames in tab links.  Clicking
this issues the update_source command.

When the update_source command's response arrives from the server, we reload the
given file's source code and refresh the corresponding tab content.  The tab
link is shaken to highlight the source code update.