Embedding the UI in the footle binary.
So far, the footle binary and the UI files have lived as separate files in the
footle distribution.  Now the UI files are bundled inside a single Go source
file and compiled as part of the footle binary.  This means the entirety of
footle can now be distributed as a single binary file.